deleted:    casting/converter.go 	modified:   pkg/config/complete_config_test.go 	modified:   pkg/elastic/indexer.go 	modified:   pkg/movedata/move.go
